+++ test/e2e/formbuilder/formbuilder.po.js
@@ -39,6 +39,7 @@ var FormBuilder = function () {
   this.questionText = element(by.id('/question/1'));
   this.questionType = element(by.id('/dataType/1'));
   this.questionAnswerText = element(by.id('/answers/text/1/1'));
+  this.prefix = element(by.id('/prefix/1'));
 
   this.basicEditTab = element(by.cssContainingText('md-tab-item span', 'Build basic properties'));
   this.advancedEditTab = element(by.cssContainingText('md-tab-item span', 'Build advanced properties'));